--- a/pkg/collector/exporter/config.go
+++ b/pkg/collector/exporter/config.go
@@ -36,6 +36,8 @@ type Config struct {
 }
 
 func (c *Config) Validate() {
+	c.Converter.Validate()
+
 	if c.Queue.MetricsBatchSize <= 0 {
 		c.Queue.MetricsBatchSize = defaultMetricsBatchSize
 	}