--- a/pkg/collector/exporter/converter/beat.go
+++ b/pkg/collector/exporter/converter/beat.go
@@ -28,6 +28,10 @@ var BeatConverter EventConverter = beatConverter{}
 
 type beatConverter struct{}
 
+func (c beatConverter) Clean() {
+	return
+}
+
 func (c beatConverter) ToEvent(token define.Token, dataId int32, data common.MapStr) define.Event {
 	return beatEvent{define.NewCommonEvent(token, dataId, data)}
 }