--- a/pkg/collector/exporter/converter/beat.go
+++ b/pkg/collector/exporter/converter/beat.go
@@ -28,9 +28,7 @@ var BeatConverter EventConverter = beatConverter{}
 
 type beatConverter struct{}
 
-func (c beatConverter) Clean() {
-	return
-}
+func (c beatConverter) Clean() {}
 
 func (c beatConverter) ToEvent(token define.Token, dataId int32, data common.MapStr) define.Event {
 	return beatEvent{define.NewCommonEvent(token, dataId, data)}