--- a/pkg/collector/exporter/converter/fta.go
+++ b/pkg/collector/exporter/converter/fta.go
@@ -27,6 +27,8 @@ var FtaConverter EventConverter = ftaConverter{}
 
 type ftaConverter struct{}
 
+func (c ftaConverter) Clean() {}
+
 func (c ftaConverter) ToEvent(token define.Token, dataId int32, data common.MapStr) define.Event {
 	return FtaEvent{define.NewCommonEvent(token, dataId, data)}
 }