--- a/pkg/collector/exporter/converter/logs.go
+++ b/pkg/collector/exporter/converter/logs.go
@@ -32,6 +32,8 @@ var LogsConverter EventConverter = logsConverter{}
 
 type logsConverter struct{}
 
+func (c logsConverter) Clean() {}
+
 func (c logsConverter) ToEvent(token define.Token, dataId int32, data common.MapStr) define.Event {
 	return logsEvent{define.NewCommonEvent(token, dataId, data)}
 }