--- a/pkg/collector/exporter/converter/proxy.go
+++ b/pkg/collector/exporter/converter/proxy.go
@@ -28,6 +28,8 @@ var ProxyConverter EventConverter = proxyConverter{}
 
 type proxyConverter struct{}
 
+func (c proxyConverter) Clean() {}
+
 func (c proxyConverter) ToEvent(token define.Token, dataId int32, data common.MapStr) define.Event {
 	return proxyEvent{define.NewCommonEvent(token, dataId, data)}
 }