#### TarsStat

```
// BenchmarkTarsStat 基准测试 TarsStat 转换性能
// goos: darwin
// goarch: arm64
// cpu: Apple M4 Pro
// Before optimization:						BenchmarkTarsStat-10    	   22473	     45333 ns/op
// After optimization:						BenchmarkTarsStat-10    	   34472	     35061 ns/op
// After optimization with Aggregate:                     BenchmarkTarsStat-10    	   27372	     43605 ns/op
```

#### TarsProperty

```
// BenchmarkTarsProperty 基准测试 TarsProperty 转换性能
// goos: darwin
// goarch: arm64
// cpu: Apple M4 Pro
// Before optimization: 			  BenchmarkTarsProperty-10    	   39517	     28596 ns/op
// After optimization:  			  BenchmarkTarsProperty-10    	   121836	     9705  ns/op
```